% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVEPLS.R
\name{BVEPLS}
\alias{BVEPLS}
\title{A basic implementation of BVE-PLS (Backward variable elimination)}
\arguments{
\item{y}{description a n*q matrix}

\item{X}{description a n*p matrix}
}
\value{
a object of class
}
\description{
A basic implementation of BVE-PLS (Backward variable elimination)
}
\examples{
y=matrix(rnorm(100),5,20)
X=matrix(rnorm(200),10,20)
fit = bve_pls(y,X)
}
