% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpSPLS.R
\name{simSPLS}
\alias{simSPLS}
\title{an improved implementation of constrained simpls}
\arguments{
\item{X}{description a n*q matrix}

\item{Y}{description a n*p matrix}
}
\value{
a list of pls components
}
\description{
an improved implementation of constrained simpls
}
\examples{
X=matrix(rnorm(200),10,20)
Y=matrix(rnorm(100),5,20)
fit = sparsePLS(X,Y,2)
}
