% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdSPLS.R
\name{svdSPLS}
\alias{svdSPLS}
\alias{sparsePLS_0}
\title{A basic implementation of constrained PLSSVD}
\usage{
sparsePLS_0(
  X,
  Y,
  h = 2,
  fullrank = TRUE,
  maxiter = 100,
  lambda1 = 0.5,
  lambda2 = 0.5
)
}
\arguments{
\item{X}{description a n*q matrix}

\item{Y}{description a n*p matrix}
}
\value{
a list of pls components
}
\description{
A basic implementation of constrained PLSSVD
}
\examples{
X=matrix(rnorm(200),10,20)
Y=matrix(rnorm(100),5,20)
fit = sparsePLS(X,Y,2)
}
