% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UVEPLS.R
\name{UVEPLS}
\alias{UVEPLS}
\alias{mcuve}
\title{Uninformative Variable Elimination in PLS}
\usage{
mcuve(X, Y, ncomp = 10, s = 30, split = 0.8, scale = TRUE)
}
\arguments{
\item{X}{The predictor matrix (numeric). If not a matrix or data.frame, it will be converted to a matrix.}

\item{Y}{The response matrix (numeric).}

\item{ncomp}{The number of components to use in the PLSR model. Default is 10.}

\item{s}{The number of iterations for variable elimination. Default is 30.}

\item{split}{The proportion of data to be used for training in each iteration. Default is 0.8.}

\item{scale}{Logical; if TRUE, the predictor matrix X will be scaled. Default is TRUE.}
}
\value{
A list containing several components of the final PLSR model: scores, loadings, Yscores, Yloadings, loading.weights, residuals, and fitted.values.
}
\description{
This function performs Multivariate Calibration by Uninformative Variable Elimination on the given datasets.
It combines the original dataset with a noise matrix, then iteratively fits a PLSR model, selecting the most informative variables and refining the model.
}
\examples{
# Example usage:
# mcuve(X = matrix(rnorm(100), 10, 10), Y = matrix(rnorm(10), 10, 1), ncomp = 15)
}
