% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIPPLS.R
\name{VIPPLS}
\alias{VIPPLS}
\alias{VIP}
\title{Calculation of variable importance in projections (VIP) scores}
\usage{
VIP(pls_object, opt.comp, num_variables = dim(pls_object$coef)[1])
}
\arguments{
\item{pls_object}{a pls model generated by "pls" package}

\item{opt.comp}{the optimal component in the pls model}

\item{num_variables}{the number of variables in the pls model; extract from the dimension of coefficients in the pls model}
}
\value{
numeric VIP scores
}
\description{
Calculation of variable importance in projections (VIP) scores
}
\examples{
pls_object = plsr(y ~ ., ncomp = opt.comp, data = data)
opt.comp = 3
num_variables = dim(pls_object$coef)[1]
Vip <- VIP(pls_object, opt.comp = opt.comp)
}
